requirements.txt
alabaster==0.7.13
annotated-types==0.5.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
Babel==2.12.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
docutils==0.18.1
hiredis==2.2.3
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.3
more-itertools==9.1.0
mypy==1.5.1
mypy-extensions==1.0.0
packaging==23.1
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
Pygments==2.16.1
python-dotenv==1.0.0
python-ulid==1.1.0
redis==4.6.0
redis-om==0.2.1
requests==2.31.0
snowballstemmer==2.2.0
Sphinx==7.1.2
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
tomli==2.0.1
types-pyOpenSSL==23.2.0.2
types-redis==4.6.0.7
typing_extensions==4.8.0
urllib3==2.0.5
